moved project_window to new cpp
created header as well
