+++ fifteen_game.h
@@ -1,1 +1,2 @@
 #include "Tile_button.h"
+#include "Project_window.h"
